Remove auto connect from useAccount
